+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## 0.0.2
+* Added analog clock UI
+* Modified readme and other documentations
+* Updated gif for demo
 
+
+## 0.0.1
 * Published on pub.dev